Changed all public properties to from int to long per docs https://developer.zendesk.com/rest_api/docs/core/introduction#id-integers
fixed #263, fixed #262

+++ src/ZendeskApi_v2/Models/HelpCenter/Categories/Category.cs
@@ -23,7 +23,7 @@ namespace ZendeskApi_v2.Models.HelpCenter.Categories
         public string HtmlUrl { get; set; }
 
         [JsonProperty("position")]
-        public int? Position { get; set; }
+        public long? Position { get; set; }
 
         [JsonProperty("created_at")]
         public string CreatedAt { get; set; }
